refactoring to allow a better user experience
major changes are being made first and foremost is the use of async and
await.
this will allow us to do some cool magic with the ui while the http
request is being loaded.
also adding some safety checks for the application so it doesn't allow
bad data for things. mainly no non alphanumeric characters in the search